move router
